+++ android/app/src/main/java/com/yagubogu/data/datasource/token/TokenDataSource.kt
@@ -1,6 +1,6 @@
-package com.yagubogu.data.datasource
+package com.yagubogu.data.datasource.token
 
-import com.yagubogu.data.dto.response.TokenResponse
+import com.yagubogu.data.dto.response.token.TokenResponse
 
 interface TokenDataSource {
     suspend fun refreshTokens(refreshToken: String): Result<TokenResponse>
