+++ android/app/src/main/java/com/yagubogu/data/network/TokenAuthenticator.kt
@@ -1,7 +1,7 @@
 package com.yagubogu.data.network
 
-import com.yagubogu.data.dto.request.TokenRequest
-import com.yagubogu.data.dto.response.TokenResponse
+import com.yagubogu.data.dto.request.token.TokenRequest
+import com.yagubogu.data.dto.response.token.TokenResponse
 import com.yagubogu.data.service.TokenApiService
 import com.yagubogu.data.util.addTokenHeader
 import com.yagubogu.data.util.getTokenFromHeader
