+++ android/app/src/main/java/com/yagubogu/data/service/TokenApiService.kt
@@ -1,7 +1,7 @@
 package com.yagubogu.data.service
 
-import com.yagubogu.data.dto.request.TokenRequest
-import com.yagubogu.data.dto.response.TokenResponse
+import com.yagubogu.data.dto.request.token.TokenRequest
+import com.yagubogu.data.dto.response.token.TokenResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
